+++ src/target/wow/WowKeywords.ts
@@ -1,5 +1,6 @@
 export enum WowKeywords {
 
     IMPORT_LIB_NAME = "__library",
-    FILE_META_VARIABLE = "__FILE_META"
+    FILE_META_VARIABLE = "__FILE_META",
+    IMPORT_EXTERNAL_MODIFIER = "__external"
 }