+++ test/target/wow/WowExports.spec.ts
@@ -24,7 +24,7 @@ import { LuaKeywords } from "../../../src/target/lua/LuaKeywords";
                 `local ${LuaKeywords.EXPORT_LOCAL_NAME} = {`,
                 `  abc = abc`,
                 `}`,
-                `${WowKeywords.IMPORT_LIB_NAME}.declare("test", ${LuaKeywords.EXPORT_LOCAL_NAME})`
+                `${WowKeywords.IMPORT_LIB_NAME}.declare("36d5a0b4f9952122e49ca2e22938ab73", ${LuaKeywords.EXPORT_LOCAL_NAME})`
             ]
         }], {
                 personalizedLibrary: false
@@ -39,10 +39,10 @@ import { LuaKeywords } from "../../../src/target/lua/LuaKeywords";
                 `local __FILE_META = {...}`,
                 `local ${LuaKeywords.EXPORT_LOCAL_NAME} = {`,
                 `}`,
-                `for mod, data in pairs(${WowKeywords.IMPORT_LIB_NAME}.get("a")) do`,
+                `for mod, data in pairs(${WowKeywords.IMPORT_LIB_NAME}.get("57a3494e7efa4eba07c02b3bdf1224e3")) do`,
                 `  ${LuaKeywords.EXPORT_LOCAL_NAME}[mod] = data`,
                 `end`,
-                `${WowKeywords.IMPORT_LIB_NAME}.declare("test", ${LuaKeywords.EXPORT_LOCAL_NAME})`
+                `${WowKeywords.IMPORT_LIB_NAME}.declare("36d5a0b4f9952122e49ca2e22938ab73", ${LuaKeywords.EXPORT_LOCAL_NAME})`
             ]
         }], {
                 personalizedLibrary: false